frontend/inventory_manager/pyproject.toml
[tool.poetry]
name = "professional-inventory-manager"
version = "1.0.0"
description = "Self-Managing Inventory Manager for sneaker/clothing shops with Lightspeed X-Series integration"
authors = ["Portfolio Project"]
readme = "README.md"
license = "Proprietary"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
# Web Framework
flask = "^3.0.0"
werkzeug = "^3.0.1"
gunicorn = "^21.2.0"
flask-cors = "^4.0.0"

# Data Processing
pandas = "^2.1.4"
numpy = "^1.24.4"

# HTTP & API
requests = "^2.31.0"
urllib3 = "^2.1.0"

# Google Sheets Integration
gspread = "^5.12.0"
google-auth = "^2.23.4"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"

# Configuration & Environment
python-dotenv = "^1.0.0"

# Background Jobs & Scheduling
apscheduler = "^3.10.4"
pytz = "^2023.3"

# Security
cryptography = ">=41.0.0,<44.0.0"

# Utilities
chardet = "^5.2.0"
simplejson = "^3.19.2"

# Database ORM (for future use)
sqlalchemy = "^2.0.23"

[tool.poetry.group.dev.dependencies]
# Testing Framework
pytest = "^8"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.21.1"
pytest-timeout = "^2"
pytest-xdist = "^3"
pytest-randomly = "^3"
freezegun = "^1"

# Linting & Formatting
ruff = "^0.6"

# Type Checking
mypy = "^1.7.1"
types-requests = "^2.31.0"
types-pytz = "^2023.3.1.1"

# Security Scanning
bandit = "^1.7.5"
safety = "^2.3.5"
pip-audit = "^2.6.1"

# Code Quality
coverage = {extras = ["toml"], version = "^7.3.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Convenient scripts (entry points)
[tool.poetry.scripts]
test = "pytest:main"
"test:fast" = "pytest:main"
lint = "ruff.__main__:main"
format = "ruff.__main__:main"

# ============================================================================
# RUFF CONFIGURATION
# ============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
indent-width = 4

# Source code directories
src = ["src", "tests"]

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "__pycache__",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # Ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "PLR0913", # Too many arguments to function call
    "PLR2004", # Magic value used in comparison
]

# Allow fix for all enabled rules (when `--fix` is provided)
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Respect magic trailing commas
skip-magic-trailing-comma = false

# Auto-detect line endings
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "S101",    # Allow assert statements in tests
    "ARG",     # Allow unused function arguments in tests
    "PLR2004", # Allow magic values in tests
]

# ============================================================================
# MYPY CONFIGURATION
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true

# Import discovery
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = true

# Output
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true

# Incremental mode
incremental = true
cache_dir = ".mypy_cache"

[[tool.mypy.overrides]]
module = [
    "gspread.*",
    "google.auth.*",
    "google.oauth2.*",
    "apscheduler.*",
    "pandas.*",
    "numpy.*",
]
ignore_missing_imports = true

# ============================================================================
# PYTEST CONFIGURATION
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q -x --maxfail=1 -n auto --dist=loadscope --timeout=10 --durations=10 --strict-markers"
asyncio_mode = "auto"

# Markers for organizing tests
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "smoke: marks tests as smoke tests",
    "e2e: marks tests as end-to-end tests",
    "no_network: marks tests that must not perform network calls",
]

# Test discovery patterns
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Coverage configuration when using pytest-cov
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
    "*/.venv/*",
    "*/scripts/*",
]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

# Coverage thresholds
fail_under = 80.0

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

# ============================================================================
# BANDIT CONFIGURATION (Security)
# ============================================================================
[tool.bandit]
exclude_dirs = [
    "tests",
    ".venv",
    "venv",
]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

# ============================================================================
# BLACK COMPATIBILITY (if needed)
# ============================================================================
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
